m491 修 m488 的构建失败 + 冒烟报告筛选器：我的 grep 漏了整整一类错误。作者回报 1.20.1 构建失败 2 个错误=int mcx = (m == dragIndex ? dragCx : nodeCx(m))，dragCx/dragCy 是 double 而 nodeCx() 返回 int，三元表达式提升成 double 再赋给 int，m488 改连线锚点时我把原来包在 sx(...) 里的表达式拆出来直接赋给 int 就炸了，改回 double；但真正的问题是这两条错误本来就在我的冒烟输出里——实测把错误放回去重跑，javac 明明白白报了 incompatible types: possible lossy conversion from double to int，它就躺在 1515 条 cannot find symbol 噪音里，而我的 grep 名单只有两类（语法错 expected/illegal/unclosed 与自家新符号 cannot find symbol），类型错误这一整类我从来没筛过；手写 grep 的问题就在这——筛的是我想得到的类别，漏的是我想不到的；docs/tools_smoke_filter.py（冒烟报告筛选器）=把 javac 输出切块，按「缺依赖三类（cannot find symbol / package 不存在 / does not override）一律记噪音，其余一律进必须看」分流，判据刻意简单可靠优先（试过按 symbol: 明细行判是不是 MC 类型，底噪仍有 476 条高到没人会看，而自家符号找不到本来就有第 19 闸悬空引用专管，在这儿混判只会互相拖累），另登记三种已知底噪（缺 MC jar 时 record 紧凑构造器、方法引用、静态导入解析不了参数类型，它们不随本仓改动增减，新出现的一律进必须看），双向验证=含错日志精确报 2 条并指到行号、修好后干净；新工具第一次跑就翻出一条我从没看见过的（SdzjzCommands 的静态导入报错，核过是缺 MC jar 的连带，该文件 m471 起没动过主线一直正常构建，登记进底噪）；这是第三次因为「验证方式」而不是「代码」出问题=m487 没筛删除的符号→第 19 闸、m488 漏带闭括号→括号配平并入第 19 闸、m491 漏筛类型错→本工具，三次的形状一模一样：我的验证是「按我记得的清单去查」而漏掉的永远是不在清单上的那类，这次改的不是清单而是把清单换成白名单式的分流（不认识的错误默认进必须看，而不是默认被忽略）；两代冒烟经筛选器复查均无必须看的真错、20 闸全绿、零配置零资源改动；教训=grep 是给已知问题用的不是给验证用的，验证要的是把未知的东西暴露出来而 grep 天然只能匹配已知模式，用它做验证等于每次只检查上次栽过的跟头；第 19 闸（静态扫描）与本筛选器（动态输出分流）两件合起来才勉强补上「我的验证」与「作者的构建」之间那道缝；版本号 0.1.491

+++ versions/1.20.1/src/main/java/com/sdzjz/retro/CanvasScreen120.java
@@ -270,7 +270,7 @@ public final class CanvasScreen120 extends AbstractContainerScreen<StructureCore
             int[] sp = g.storageNodePos.get(e[1]);
             if (m >= g.machineNodes.size() || sp == null) continue;
             // m488：锚点按 100×52 卡（出口柱在右缘中上、进口柱在右缘中下；存储卡仍是 24×24）
-            int mcx = (m == dragIndex ? dragCx : nodeCx(m)), mcy = (m == dragIndex ? dragCy : nodeCy(m));
+            double mcx = (m == dragIndex ? dragCx : nodeCx(m)), mcy = (m == dragIndex ? dragCy : nodeCy(m)); // 拖动位是 double
             boolean 产出 = e[2] == 0;
             double x1 = sx(mcx + NW), y1 = sy(mcy + (产出 ? NH / 2 - 7 : NH / 2 + 7));
             double x2 = sx(stX(e[1], sp) + 12), y2 = sy(stY(e[1], sp) + 12);
